Debugging

+++ baselines/pfsp/lib/Pool_ext.c
@@ -42,7 +42,7 @@ void pushBackBulk(SinglePool_ext* pool, Node* nodes, int size) {
     expected = false;
     if (atomic_compare_exchange_strong(&(pool->lock), &expected, true)) {
       if (pool->front + pool->size + size >= pool->capacity) {
-	pool->capacity *= pow(2,ceil(log2((pool->front + pool->size + size) / pool->capacity))); //(size+pool->front+1);
+	pool->capacity *= pow(2,ceil(log2((double)(pool->front + pool->size + size) / pool->capacity))); //(size+pool->front+1);
 	pool->elements = realloc(pool->elements, pool->capacity * sizeof(Node));
 	printf("\nRealloc: PushBackBulk with k = %f \n",log2((double)(pool->front + pool->size + size) / pool->capacity));
       }
